changes made to execution and built-in
