+++ 05_Execution/single_command_utils7.c
@@ -13,66 +13,8 @@
 #include "../includes/minishell.h"
 
 /**
- * @function: handle_parent_for_handling_forking_process
- * @brief: handle parent process for the function handling forking process below
- * 
- * @param t_redirect_single_command_params *params: structure for
- 	single_command parameters
- * 
- * @return: void function
- */
-
-void	handle_parent_for_handling_forking_process(
-			t_redirect_single_command_params *params)
-{
-	int	status;
-
-	ignore_parent_signals();
-	waitpid(params->pid, &status, 0);
-	if (WIFEXITED(status))
-	{
-		ft_dprintf(1, "Child %d exited normally, with exit code %d\n", params->pid, WEXITSTATUS(status));
-		*params->exit_status = WEXITSTATUS(status);
-	}
-	else if (WIFSIGNALED(status))
-	{
-		ft_dprintf(1, "Child %d exited with signals, with exit code %d\n", params->pid, WTERMSIG(status));
-		*params->exit_status = WTERMSIG(status) + 128;
-	}
-	ft_dprintf(1, "Current exit status %d\n", *params->exit_status);
-	ft_signal(NULL, NULL, NULL, PARENT);
-}
-
-/**
- * @function: handle_delimiter_input_single_commands
- * @brief: checking if expanstion is necessary if heredocs is met
- 	and cleaning input after that
- 	0 means not necessary. 1 means necessary
- * 
- * @param t_redirect_single_command_params *params: structure for
- 	single_command parameters
- * 
- * @return: void function
- */
-
-void	handle_delimiter_input_single_commands(
-			t_redirect_single_command_params *params)
-{
-	params->ignore_quote = 0;
-	if (ft_has_quote(params->result->delimiter[params->delimiter_counter]) == 1)
-		params->ignore_quote = 0;
-	else
-		params->ignore_quote = 1;
-	params->result->delimiter[params->delimiter_counter]
-		= ft_remove_quote(params->result->delimiter[params->delimiter_counter]);
-	ft_dprintf(2, "current delimiter now after clean up %s\n",
-		params->result->delimiter[params->delimiter_counter]);
-	ft_dprintf(2, "value of params ignore_quote [%d]\n", params->ignore_quote);
-}
-
-/**
- * @function: handling_forking_process
- * @brief: handles the heredoc process once "<<" is found in commands
+ * @function: heredocs
+ * @brief: handling the presence of heredocs
  * 
  * @param t_redirect_single_command_params *params: structure for
  	single_command parameters
@@ -82,81 +24,29 @@ void	handle_delimiter_input_single_commands(
  * @return: -1 if failure, 0 if success
  */
 
-int	handling_forking_process(
-			t_redirect_single_command_params *params, char ***env)
+int	heredocs(t_redirect_single_command_params *params, char ***env)
 {
-	ft_dprintf(2, "Welcome to heredocs <<\n");
-	ft_dprintf(2, "delimiter for heredocs: %s\n",
-		params->result->delimiter[params->delimiter_counter]);
-	if (*params->exit_status != 0)
-		return (-1);
-	handle_delimiter_input_single_commands(params);
-	params->pid = fork();
-	if (params->pid < 0)
+	params->a = 0;
+	while (params->result->redirect[params->x] != NULL)
 	{
-		perror("fork failed");
-		return (-1);
-	}
-	if (params->pid == 0)
-	{
-		ft_signal(params, NULL, *env, CHILD);
-		handle_heredoc_child_process(params, env);
-	}
-	else
-		handle_parent_for_handling_forking_process(params);
-	params->delimiter_counter++;
-	params->z++;
-	closing_current_pipe_after_writing_data(params);
-	return (0);
-}
-
-/**
- * @function: handling_next_redirect
- * @brief: looping through the redirect array to find the next redirect
- 	changing the redirect array if "<<" is found to another char
- * 
- * @param t_redirect_single_command_params *params: structure for
- 	single_command parameters
- * 
- * @return: void function
- */
-
-void	handling_next_redirect(t_redirect_single_command_params *params)
-{	
-	ft_dprintf(2, "Entering here to change current redirect array %s\n",
-		params->result->redirect[params->x]);
-	free(params->result->redirect[params->x]);
-	params->result->redirect[params->x] = ft_strdup("a");
-	ft_dprintf(2, "After changing redirect array %s\n",
-		params->result->redirect[params->x]);
-	params->x++;
-}
-
-/**
- * @function: handling_last_redirect
- * @brief: handling the last_redirect and checking if there is a command
- * 
- * @param t_redirect_single_command_params *params: structure for
- 	single_command parameters
- 	***env: *** is called in the calling function
- 	needed ** to free data if child process exits.
- * 
- * @return: void function
- */
-
-int	handling_last_redirect(
-			t_redirect_single_command_params *params, char ***env)
-{
-	ft_dprintf(2, "Entering here to change current redirect array %s\n",
-		params->result->redirect[params->x]);
-	free(params->result->redirect[params->x]);
-	params->result->redirect[params->x] = ft_strdup("a");
-	ft_dprintf(2, "After changing redirect array %s\n",
-		params->result->redirect[params->x]);
-	if (params->result->cmd[0] != NULL)
-	{
-		if (handle_child_execution(params, env) == -1)
-			return (-1);
+		if ((ft_strcmp(params->result->redirect[params->x], "<<") == 0))
+		{
+			if (handling_forking_process(params, env) == -1)
+				return (-1);
+			if (params->result->redirect[params->x + 1] != NULL)
+			{
+				ft_dprintf(2, "Debugging next redirect\n");
+				handling_next_redirect(params);
+				continue ;
+			}
+			else if (params->result->redirect[params->x + 1] == NULL)
+			{
+				ft_dprintf(2, "Debugging handling last redirect <<\n");
+				if (handling_last_redirect(params, env) == -1)
+					return (-1);
+			}
+		}
+		params->x++;
 	}
 	return (0);
 }
